+++ src/bin/cli.js
@@ -229,6 +229,12 @@ function printUsage() {
     '  --image <image>           Docker image (required for docker isolation)'
   );
   console.log('  --user <username>         Run command as specified user');
+  console.log(
+    '  --keep-alive, -k          Keep isolation environment alive after command exits'
+  );
+  console.log(
+    '  --auto-remove-docker-container  Automatically remove docker container after exit (disabled by default)'
+  );
   console.log('  --version, -v             Show version information');
   console.log('');
   console.log('Examples:');
@@ -368,6 +374,8 @@ async function runWithIsolation(options, cmd) {
     image: options.image,
     detached: mode === 'detached',
     user: options.user,
+    keepAlive: options.keepAlive,
+    autoRemoveDockerContainer: options.autoRemoveDockerContainer,
   });
 
   // Get exit code